Fix classpath jar assuming all entries are jars
This commit fixes an assumption when building the classpath jar,
that all elements on classpath are jars themselves. It is possible
that the classpath also consists of directories.

Fixes #152
